+++ src/main/java/org/ktc2/cokaen/wouldyouin/auth/MemberIdentifier.java
@@ -1,4 +1,4 @@
-package org.ktc2.cokaen.wouldyouin.security;
+package org.ktc2.cokaen.wouldyouin.auth;
 
 import org.ktc2.cokaen.wouldyouin.member.persist.MemberType;
