+++ src/main/java/org/ktc2/cokaen/wouldyouin/member/api/HostController.java
@@ -6,8 +6,8 @@ import org.ktc2.cokaen.wouldyouin.member.application.dto.HostEditRequest;
 import org.ktc2.cokaen.wouldyouin.member.application.dto.MemberResponse;
 import org.ktc2.cokaen.wouldyouin.member.persist.MemberType;
 import org.ktc2.cokaen.wouldyouin._common.api.ApiResponseBody;
-import org.ktc2.cokaen.wouldyouin.security.Authorize;
-import org.ktc2.cokaen.wouldyouin.security.MemberIdentifier;
+import org.ktc2.cokaen.wouldyouin.auth.Authorize;
+import org.ktc2.cokaen.wouldyouin.auth.MemberIdentifier;
 import org.ktc2.cokaen.wouldyouin.member.application.HostService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
