+++ src/main/java/org/ktc2/cokaen/wouldyouin/member/api/MemberController.java
@@ -6,8 +6,8 @@ import org.ktc2.cokaen.wouldyouin.member.application.dto.MemberEditRequest;
 import org.ktc2.cokaen.wouldyouin.member.application.dto.MemberResponse;
 import org.ktc2.cokaen.wouldyouin.member.persist.MemberType;
 import org.ktc2.cokaen.wouldyouin._common.api.ApiResponseBody;
-import org.ktc2.cokaen.wouldyouin.security.Authorize;
-import org.ktc2.cokaen.wouldyouin.security.MemberIdentifier;
+import org.ktc2.cokaen.wouldyouin.auth.Authorize;
+import org.ktc2.cokaen.wouldyouin.auth.MemberIdentifier;
 import org.ktc2.cokaen.wouldyouin.member.application.MemberDeleteService;
 import org.ktc2.cokaen.wouldyouin.member.application.MemberSearchService;
 import org.ktc2.cokaen.wouldyouin.member.application.MemberService;
